+++ test/unit/rmq_metadata/summarize_data.py
@@ -73,7 +73,7 @@ class UnitTest(unittest.TestCase):
         self.results2 = [(u'London', u'LOCATION'), (u'SW1W9AX', u'LOCATION')]
 
     @mock.patch("rmq_metadata.merge_data")
-    @mock.patch("rmq_metadata._sort_data")
+    @mock.patch("rmq_metadata.sort_data")
     def test_end_loop_data(self, mock_sort, mock_merge):
 
         """Function:  test_end_loop_data
@@ -88,8 +88,9 @@ class UnitTest(unittest.TestCase):
                                  self.loop4a]
         mock_merge.return_value = self.data_list
 
-        self.assertEqual(rmq_metadata.summarize_data(
-            self.categorized_text2, self.token_types), self.results2)
+        self.assertEqual(
+            rmq_metadata.summarize_data(
+                self.categorized_text2, self.token_types), self.results2)
 
     def test_empty_categorized_text(self):
 
@@ -103,7 +104,7 @@ class UnitTest(unittest.TestCase):
 
         self.assertEqual(rmq_metadata.summarize_data([], self.token_types), [])
 
-    @mock.patch("rmq_metadata._sort_data")
+    @mock.patch("rmq_metadata.sort_data")
     def test_summarize_data(self, mock_sort):
 
         """Function:  test_summarize_data
@@ -117,8 +118,9 @@ class UnitTest(unittest.TestCase):
         mock_sort.side_effect = [self.loop1, self.loop2, self.loop3,
                                  self.loop4]
 
-        self.assertEqual(rmq_metadata.summarize_data(
-            self.categorized_text, self.token_types), self.results)
+        self.assertEqual(
+            rmq_metadata.summarize_data(
+                self.categorized_text, self.token_types), self.results)
 
 
 if __name__ == "__main__":