+++ test/unit/rmq_metadata/unit_test_run.sh
@@ -6,10 +6,7 @@
 
 echo ""
 echo "Unit test:  rmq_metadata.py"
-test/unit/rmq_metadata/_convert_data.py
-test/unit/rmq_metadata/_process_queue.py
-test/unit/rmq_metadata/_sort_data.py
-test/unit/rmq_metadata/_validate_files.py
+test/unit/rmq_metadata/convert_data.py
 test/unit/rmq_metadata/create_metadata.py
 test/unit/rmq_metadata/extract_pdf.py
 test/unit/rmq_metadata/find_tokens.py
@@ -22,9 +19,12 @@ test/unit/rmq_metadata/merge_data.py
 test/unit/rmq_metadata/monitor_queue.py
 test/unit/rmq_metadata/non_proc_msg.py
 test/unit/rmq_metadata/pdf_to_string.py
+test/unit/rmq_metadata/process_message.py
 test/unit/rmq_metadata/process_msg.py
 test/unit/rmq_metadata/read_pdf.py
 test/unit/rmq_metadata/run_program.py
+test/unit/rmq_metadata/sort_data.py
 test/unit/rmq_metadata/summarize_data.py
 test/unit/rmq_metadata/validate_create_settings.py
+test/unit/rmq_metadata/validate_files.py